New parser implementation (incomplete)
Ongoing work to create a new streamlined token parser.  The main
motivation here is to consolidate a lot of convoluted logic of the
current parser which exists mainly to support a kind of lookahead.

This new implementation will use

Only the basics are currently implemented:
- Group tokens are successfully gathered
- Most variable cases are gathered, with index separate from value
- Derived types successfully create subdicts (although index is ignored)

What is not done,
- Variable content is not parsed, tokens are gathered into a list
- Index content is also not parsed, and is gathered into a string
- Patching is completely disabled, whitespace and comments are discarded

There is a lot to do here, but current work is promising.