New parser: Value parsing
Continuing progress on the new parser.

I have recycled most of the `_parse_value` function and implemented a
`read_value` function.  Unlike its predecessor, this constructs a

This only parses the content after the assignment operator (=), it
doesn't try to properly index the vector or merge with existing data.

Tricky cases like complex variables and unassigned elements with
repeated commas (including the tail-comma case) appear to be working
correctly.  But not everything has been tested.