Derived type dict updates
The parser now updates components of existing derived types, instead of
overwriting them.  It works by parsing each entry as a derived type (or
sequence of them) and then iteratively applies this to the existing
entry.

This might not play well with vectors of derived types, but it appears
to work on exisiting tests.